leetcode/1268.Search-Suggestions-System/1268.SearchSuggestionsSystem_test.go
package leetcode

import (
	"reflect"
	"testing"
)

func Test_suggestedProducts(t *testing.T) {
	type args struct {
		products   []string
		searchWord string
	}
	tests := []struct {
		name string
		args args
		want [][]string
	}{
		{
			name: `products = ["mobile","mouse","moneypot","monitor","mousepad"], searchWord = "mouse"`,
			args: args{[]string{"mobile", "mouse", "moneypot", "monitor", "mousepad"}, "mouse"},
			want: [][]string{
				{"mobile", "moneypot", "monitor"},
				{"mobile", "moneypot", "monitor"},
				{"mouse", "mousepad"},
				{"mouse", "mousepad"},
				{"mouse", "mousepad"},
			},
		},
		{
			name: "havana",
			args: args{[]string{"havana"}, "tatiana"},
			want: make([][]string, 7),
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := suggestedProducts(tt.args.products, tt.args.searchWord); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("suggestedProducts() = %v, want %v", got, tt.want)
			}
		})
	}
}
